requirements.txt
pandas>=1.5.0
requests>=2.28.0
sqlalchemy>=2.0.0
python-dateutil>=2.8.2
pymysql>=1.1.0
cryptography>=41.0.0
python-dotenv>=1.0.0 